Phase 1: Extract app_paths.py from cli.py
Moves CURRENT_DIRECTORY constant, _legacy_path_warnings tracking set,
resolve_config_path, and log_legacy_path_warnings out of cli.py and
into a focused app_paths module.

cli.py imports from app_paths so existing call sites that do
'cli.resolve_config_path(...)' continue to work via Python's
module-level re-export semantics.

Test updates: test_cli_dispatch.py monkeypatches now target
app_paths.CURRENT_DIRECTORY and app_paths._legacy_path_warnings (reading
inside resolve_config_path resolves them in the function's defining
module, not via cli). test_config_resolution.py is unchanged: its
'from rushti.cli import resolve_config_path' still works because cli
imports the function at module scope.
